-캠페인 상세 모달창 추가, 패키지 이름 변경(kh->gnt)
