+++ gibuAndTakePrj/src/com/gnt/campaign/controller/CampaignDetailServlet.java
@@ -1,4 +1,4 @@
-package com.kh.gibu.controller;
+package com.gnt.campaign.controller;
 
 import java.io.IOException;
 
@@ -8,12 +8,11 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet(urlPatterns = "/view/gibu")
-public class GibuServlet extends HttpServlet{
+@WebServlet(urlPatterns = "/view/campaign_detail")
+public class CampaignDetailServlet extends HttpServlet{
 
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/views/select_gibu/gibu_ing.jsp").forward(req, resp);
+		req.getRequestDispatcher("/views/select_campaign/campaign_detail.jsp").forward(req, resp);
 	}
-	
 }
