+++ gibuAndTakePrj/src/com/gnt/funding/controller/FundingDetailServlet.java
@@ -1,4 +1,4 @@
-package com.kh.gibu.controller;
+package com.gnt.funding.controller;
 
 import java.io.IOException;
 
@@ -8,12 +8,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet(urlPatterns = "/view/gibu_done")
-public class GibuDoneServlet extends HttpServlet{
-	
-	
+@WebServlet(urlPatterns = "/view/funding_detail")
+public class FundingDetailServlet extends HttpServlet{
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/views/select_gibu/gibu_finish.jsp").forward(req, resp);
+		req.getRequestDispatcher("/views/select_funding/funding_detail.jsp").forward(req, resp);
 	}
 }
