+++ gibuAndTakePrj/src/com/gnt/funding/controller/FundingDoneServlet.java
@@ -1,4 +1,4 @@
-package com.kh.campaign.controller;
+package com.gnt.funding.controller;
 
 import java.io.IOException;
 
@@ -8,11 +8,11 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet(urlPatterns = "/view/campaign_detail")
-public class CampaignDetailServlet extends HttpServlet{
+@WebServlet(urlPatterns = "/view/funding_done")
+public class FundingDoneServlet extends HttpServlet {
 
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/views/select_campaign/campaign_detail.jsp").forward(req, resp);
+		req.getRequestDispatcher("/views/select_funding/funding_finish.jsp").forward(req, resp);
 	}
 }
