+++ gibuAndTakePrj/src/com/gnt/funding/controller/FundingServlet.java
@@ -1,4 +1,4 @@
-package com.kh.gibu.controller;
+package com.gnt.funding.controller;
 
 import java.io.IOException;
 
@@ -8,11 +8,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet(urlPatterns = "/view/gibu_detail")
-public class GibuDetailServlet extends HttpServlet {
-	
+@WebServlet(urlPatterns = "/view/funding")
+public class FundingServlet extends HttpServlet{
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/views/select_gibu/gibu_detail.jsp").forward(req, resp);
+		req.getRequestDispatcher("/views/select_funding/funding_ing.jsp").forward(req, resp);
 	}
 }
