+++ gibuAndTakePrj/src/com/gnt/gibu/controller/GibuDetailServlet.java
@@ -1,4 +1,4 @@
-package com.kh.funding.controller;
+package com.gnt.gibu.controller;
 
 import java.io.IOException;
 
@@ -8,11 +8,11 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet(urlPatterns = "/view/funding_done")
-public class FundingDoneServlet extends HttpServlet {
-
+@WebServlet(urlPatterns = "/view/gibu_detail")
+public class GibuDetailServlet extends HttpServlet {
+	
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/views/select_funding/funding_finish.jsp").forward(req, resp);
+		req.getRequestDispatcher("/views/select_gibu/gibu_detail.jsp").forward(req, resp);
 	}
 }
