+++ gibuAndTakePrj/src/com/gnt/gibu/controller/GibuServlet.java
@@ -1,4 +1,4 @@
-package com.kh.funding.controller;
+package com.gnt.gibu.controller;
 
 import java.io.IOException;
 
@@ -8,10 +8,12 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet(urlPatterns = "/view/funding_detail")
-public class FundingDetailServlet extends HttpServlet{
+@WebServlet(urlPatterns = "/view/gibu")
+public class GibuServlet extends HttpServlet{
+
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/views/select_funding/funding_detail.jsp").forward(req, resp);
+		req.getRequestDispatcher("/views/select_gibu/gibu_ing.jsp").forward(req, resp);
 	}
+	
 }
